+++ SOURCE_FILES/Adapt_Mesh.cuh
@@ -43,7 +43,7 @@ __global__ void D__Wavelet_Transform__F(T* 					PDF,
 
 		// find the dyadic cube indeces and do the wavelet transform at the same time
 		#pragma unroll
-		for(uint16_t d = 0; d < DIMENSIONS; d++){
+		for(uint32_t d = 0; d < DIMENSIONS; d++){
 			for(uint32_t k = 0; k < powf(2,DIMENSIONS);k++){
 				if( floor(positive_rem(k, pow(2, d + 1)) / pow(2, d))  == 0 ){ // this part decides whether this is an approximation node
 					uint32_t approx_coef_idx = main_node_idx;
@@ -91,11 +91,11 @@ int16_t ADAPT_MESH_REFINEMENT_nD(const thrust::host_vector<TYPE>& H_PDF, thrust:
 	thrust::host_vector<AMR_node_select> 	H__Node_selection(Total_Points, {0,0});
 	thrust::device_vector<AMR_node_select>	D__Node_selection = H__Node_selection;
 
-	for (uint16_t k = 0; k < LvlFine - LvlCoarse; k++) {
+	for (uint32_t k = 0; k < LvlFine - LvlCoarse; k++) {
 
 		uint32_t Points_at_level  = PtsPerDim / rescaling;
 
-		uint16_t Threads = fmin(THREADS_P_BLK, pow(Points_at_level, DIMENSIONS));
+		uint32_t Threads = fmin(THREADS_P_BLK, pow(Points_at_level, DIMENSIONS));
 		uint32_t Blocks  = floor((pow(Points_at_level, DIMENSIONS) - 1) / Threads) + 1;
 
 		D__Wavelet_Transform__F<TYPE> <<<Blocks, Threads>>> (rpc((*D__PDF),0), rpc(D__Node_selection,0),PtsPerDim,Points_at_level,rescaling);
